+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 +## [unreleased]
 
+- feat(ninjs): output profile and slugline in ninjs
 - fix(celery): use application context for task error handling
 - fix(websockets): remove client when connection is closed
 - add `service.find` method which is using mongo syntax for queries